Reduce disk requirement for half density on SP216
In previous versions of this document, the disk requirement for indexing 10 M items in SP2016 was 300 GB.  For some reason the latest version has it listed as 500 GB - the same amount of storage needed to index 20 M items.  This makes no sense, so my best guess is that this change was inadvertent.